sui_verifier_latest

Module private_generics_verifier_v2

Constants

COIN_REGISTRY_MODULE

EVENT_MODULE

EXHAUSTIVE_MODULES

FUNCTIONS_TO_CHECK

INTERNAL_MODULE

MOVE_STDLIB_INTERNAL_PERMIT

SUI_COIN_REGISTRY_NEW_CURRENCY

SUI_EVENT_EMIT_AUTHENTICATED

SUI_EVENT_EMIT_EVENT

SUI_EVENT_EVENTS_BY_TYPE

SUI_EVENT_NUM_EVENTS

SUI_TRANSFER_FREEZE_OBJECT

SUI_TRANSFER_PARTY_TRANSFER

SUI_TRANSFER_PUBLIC_FREEZE_OBJECT

SUI_TRANSFER_PUBLIC_PARTY_TRANSFER

SUI_TRANSFER_PUBLIC_RECEIVE

SUI_TRANSFER_PUBLIC_SHARE_OBJECT

SUI_TRANSFER_PUBLIC_TRANSFER

SUI_TRANSFER_RECEIVE

SUI_TRANSFER_RECEIVING_OBJECT_ID

SUI_TRANSFER_SHARE_OBJECT

SUI_TRANSFER_TRANSFER

TRANSFER_MODULE

Functions

callee_package_name

help_message

verify_module

Several functions in the Sui Framework have internal type parameters, whose arguments must be instantiated with types defined in the caller’s module. For example, with transfer::transfer<T>(...) T must be a type declared in the current module. Otherwise, transfer::public_transfer<T>(...) can be used without restriction, as long as T has store. Note thought that the ability constraint is not checked in this verifier, but rather in the normal bytecode verifier type checking. To avoid, issues, all su::transfer and sui::event functions must be configured in INTERNAL_FUNCTIONS.